Renamespace CallbackHttpErrorEvent
