+++ tests/Functional/MessageDispatcher/SendCallbackMessageDispatcherTest.php
@@ -8,8 +8,8 @@ use App\Entity\Callback\CallbackInterface;
 use App\Entity\Callback\CompileFailureCallback;
 use App\Entity\Callback\DelayedCallback;
 use App\Entity\Callback\ExecuteDocumentReceivedCallback;
-use App\Event\Callback\CallbackHttpErrorEvent;
 use App\Event\CallbackEventInterface;
+use App\Event\CallbackHttpErrorEvent;
 use App\Event\SourceCompile\SourceCompileFailureEvent;
 use App\Event\TestExecuteDocumentReceivedEvent;
 use App\Message\SendCallback;
